+++ CHANGELOG.md
@@ -44,6 +44,12 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
 - Added `--label-columns` / `-L` flag to present labels as columns.
   (#55, thanks @semihbkgr!)
 
+- Added parsing of a pod's "RESTART" column (e.g `5 (3m ago)`)
+  so it auto updates, similarly to the "AGE" column. (#56)
+
+- Added timer on pod's "STATUS" column when a pod is deleted
+  (e.g `Deleted (3m ago)`). (#56)
+
 ## v0.4.0 (2023-09-03)
 
 - Added text filtering. (#32, thanks @semihbkgr!)